Fix opentelemetry test (#2043)
